Switched numbers from float to double.

+++ interpreter/alien_dl.c
@@ -1,6 +1,6 @@
 /*
  * nix operating system project tre interpreter
- * Copyright (c) 2005-2007 Sven Klose <pixel@copei.de
+ * Copyright (c) 2005-2008 Sven Klose <pixel@copei.de
  *
  * Dynamic linker support
  */
@@ -20,8 +20,8 @@
 treptr
 trealien_builtin_dlopen (treptr args)
 {
-    treptr path = trearg_get (args);
-    void *hdl;
+    treptr  path = trearg_get (args);
+    void    * hdl;
 
     while (TREPTR_IS_STRING(path) == FALSE)
         path = treerror (path, "path to shared object expected");
@@ -29,23 +29,23 @@ trealien_builtin_dlopen (treptr args)
     hdl = dlopen (TREATOM_STRINGP(path), RTLD_NOW);
     if (hdl == NULL)
         return treerror (path, dlerror ());
-    return treatom_number_get ((float) (int) hdl, TRENUMTYPE_INTEGER);
+    return treatom_number_get ((double) (int) hdl, TRENUMTYPE_INTEGER);
 }
 
 /* Close shared object. */
 treptr
 trealien_builtin_dlclose (treptr args)
 {
-    treptr hdl = trearg_get (args);
+    treptr  hdl = trearg_get (args);
     int     ret;
 
     while (TREPTR_IS_NUMBER(hdl) == FALSE)
         hdl = treerror (hdl, "number handle expected");
 
-    ret = dlclose ((void*) (int) TRENUMBER_VAL(hdl));
+    ret = dlclose ((void *) (int) TRENUMBER_VAL(hdl));
     if (ret == -1)
         return treerror (hdl, dlerror ());
-    return treatom_number_get ((float) ret, TRENUMTYPE_INTEGER);
+    return treatom_number_get ((double) ret, TRENUMTYPE_INTEGER);
 }
 
 /* Get pointer to symbol. */
@@ -54,7 +54,7 @@ trealien_builtin_dlsym (treptr args)
 {
     treptr  hdl;
     treptr  sym;
-    void     *ret;
+    void    * ret;
 
     trearg_get2 (&hdl, &sym, args);
 
@@ -64,10 +64,10 @@ trealien_builtin_dlsym (treptr args)
     while (TREPTR_IS_STRING(sym) == FALSE)
         sym = treerror (sym, "symbol string expected");
 
-    ret = dlsym ((void*) (int) TRENUMBER_VAL(hdl), TREATOM_STRINGP(sym));
+    ret = dlsym ((void *) (int) TRENUMBER_VAL(hdl), TREATOM_STRINGP(sym));
     if (ret == NULL)
         return treerror (hdl, dlerror ());
-    return treatom_number_get ((float) (int) ret, TRENUMTYPE_INTEGER);
+    return treatom_number_get ((double) (int) ret, TRENUMTYPE_INTEGER);
 }
 
 size_t
@@ -88,27 +88,27 @@ trealien_argconv (treptr arg)
 treptr
 trealien_builtin_dlcall0 (treptr args)
 {
-    int      ret;
+    int     ret;
     treptr  ptr;
-    int      (*fun) (void);
+    int     (* fun) (void);
 
     ptr = trearg_get (args);
     while (TREPTR_IS_NUMBER(ptr) == FALSE)
         ptr = treerror (ptr, "number expected");
 
     fun = (void *) (int) TRENUMBER_VAL(ptr);
-    ret = (*fun) ();
-    return treatom_number_get ((float) ret, TRENUMTYPE_INTEGER);
+    ret = (* fun) ();
+    return treatom_number_get ((double) ret, TRENUMTYPE_INTEGER);
 }
 
 /* Call C function with 1 arguments. */
 treptr
 trealien_builtin_dlcall1 (treptr args)
 {
-    int      ret;
+    int     ret;
     treptr  ptr;
-    int      (*fun) (int);
-    int      a1;
+    int     (* fun) (int);
+    int     a1;
 
     ptr = CAR(args);
     while (TREPTR_IS_NUMBER(ptr) == FALSE)
@@ -117,20 +117,20 @@ trealien_builtin_dlcall1 (treptr args)
     a1 = trealien_argconv (CADR(args));
 
     fun = (void *) (int) TRENUMBER_VAL(ptr);
-    ret = (*fun) (a1);
-    return treatom_number_get ((float) ret, TRENUMTYPE_INTEGER);
+    ret = (* fun) (a1);
+    return treatom_number_get ((double) ret, TRENUMTYPE_INTEGER);
 }
 
 /* Call C function with 2 arguments. */
 treptr
 trealien_builtin_dlcall2 (treptr args)
 {
-    int      ret;
+    int     ret;
     treptr  ptr;
     treptr  a;
-    int      (*fun) (int, int);
-    int      a1;
-    int      a2;
+    int     (* fun) (int, int);
+    int     a1;
+    int     a2;
 
     ptr = CAR(args);
     while (TREPTR_IS_NUMBER(ptr) == FALSE)
@@ -142,21 +142,21 @@ trealien_builtin_dlcall2 (treptr args)
     a2 = trealien_argconv (CAR(a));
 
     fun = (void *) (int) TRENUMBER_VAL(ptr);
-    ret = (*fun) (a1, a2);
-    return treatom_number_get ((float) ret, TRENUMTYPE_INTEGER);
+    ret = (* fun) (a1, a2);
+    return treatom_number_get ((double) ret, TRENUMTYPE_INTEGER);
 }
 
 /* Call C function with 3 arguments. */
 treptr
 trealien_builtin_dlcall3 (treptr args)
 {
-    int      ret;
+    int     ret;
     treptr  ptr;
     treptr  a;
-    int      (*fun) (int, int, int);
-    int      a1;
-    int      a2;
-    int      a3;
+    int     (* fun) (int, int, int);
+    int     a1;
+    int     a2;
+    int     a3;
 
     ptr = CAR(args);
     while (TREPTR_IS_NUMBER(ptr) == FALSE)
@@ -170,22 +170,22 @@ trealien_builtin_dlcall3 (treptr args)
     a3 = trealien_argconv (CAR(a));
 
     fun = (void *) (int) TRENUMBER_VAL(ptr);
-    ret = (*fun) (a1, a2, a3);
-    return treatom_number_get ((float) ret, TRENUMTYPE_INTEGER);
+    ret = (* fun) (a1, a2, a3);
+    return treatom_number_get ((double) ret, TRENUMTYPE_INTEGER);
 }
 
 /* Call C function with 4 arguments. */
 treptr
 trealien_builtin_dlcall4 (treptr args)
 {
-    int      ret;
+    int     ret;
     treptr  ptr;
     treptr  a;
-    int      (*fun) (int, int, int, int);
-    int      a1;
-    int      a2;
-    int      a3;
-    int      a4;
+    int     (* fun) (int, int, int, int);
+    int     a1;
+    int     a2;
+    int     a3;
+    int     a4;
 
     ptr = CAR(args);
     while (TREPTR_IS_NUMBER(ptr) == FALSE)
@@ -201,6 +201,6 @@ trealien_builtin_dlcall4 (treptr args)
     a4 = trealien_argconv (CAR(a));
 
     fun = (void *) (int) TRENUMBER_VAL(ptr);
-    ret = (*fun) (a1, a2, a3, a4);
-    return treatom_number_get ((float) ret, TRENUMTYPE_INTEGER);
+    ret = (* fun) (a1, a2, a3, a4);
+    return treatom_number_get ((double) ret, TRENUMTYPE_INTEGER);
 }
